requirements.txt
PySide6>=6.5.0
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
qtawesome>=1.3.0
